refactor: simplify conflicts page state and detail views
