+++ src/pages/conflicts/ConflictsPage.tsx
@@ -1,20 +1,9 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSearchParams } from 'react-router-dom';
-import { getConflicts, updateConflict } from '@psycron/api/user/conflicts';
-import type {
-	ConflictStatus,
-	ConflictType,
-	IConflict,
-} from '@psycron/api/user/conflicts/index.types';
-import { useAlert } from '@psycron/context/alert/AlertContext';
-import { useTherapistId } from '@psycron/hooks/useTherapistId';
 import { PageLayout } from '@psycron/layouts/app/pages-layout/PageLayout';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { format } from 'date-fns';
 
 import { ConflictDetail } from './components/conflict-detail/ConflictDetail';
-import { mockedConflicts } from './ConflictsPage.mocks';
+import { useConflictsPageState } from './hooks/useConflictsPageState';
 import {
 	ConflictCard,
 	ConflictCardDate,
@@ -30,8 +19,6 @@ import {
 	FiltersLabel,
 	FiltersRow,
 	FiltersSection,
-	MockModeNotice,
-	MockModeText,
 	SidebarCount,
 	SidebarEyebrow,
 	SidebarHeader,
@@ -46,113 +33,25 @@ import {
 
 export const ConflictsPage = () => {
 	const { t } = useTranslation();
-	const [searchParams] = useSearchParams();
-	const therapistId = useTherapistId();
-	const { showAlert } = useAlert();
-	const queryClient = useQueryClient();
-	const explicitMockMode = searchParams.get('debugConflicts') === 'mock';
-	const forceLiveMode = searchParams.get('debugConflicts') === 'live';
-
-	const [statusFilter, setStatusFilter] = useState<ConflictStatus | undefined>(
-		'OPEN'
-	);
-	const [typeFilter, setTypeFilter] = useState<ConflictType | undefined>();
-	const [selectedConflictId, setSelectedConflictId] = useState<string | null>(
-		null
-	);
-
-	const { data, isLoading } = useQuery({
-		queryKey: ['conflicts', therapistId, statusFilter, typeFilter],
-		queryFn: () =>
-			getConflicts({ status: statusFilter, therapistId, type: typeFilter }),
-		enabled: Boolean(therapistId) && !explicitMockMode,
-	});
-
-	const useMockConflicts =
-		import.meta.env.DEV &&
-		(explicitMockMode ||
-			(!forceLiveMode && !isLoading && (data?.conflicts?.length ?? 0) === 0));
-
-	const conflicts = useMemo(() => {
-		const sourceConflicts = useMockConflicts
-			? mockedConflicts
-			: (data?.conflicts ?? []);
-
-		return sourceConflicts.filter((conflict: IConflict) => {
-			const matchesStatus = !statusFilter || conflict.status === statusFilter;
-			const matchesType = !typeFilter || conflict.type === typeFilter;
-
-			return matchesStatus && matchesType;
-		});
-	}, [data?.conflicts, statusFilter, typeFilter, useMockConflicts]);
-
-	useEffect(() => {
-		if (!conflicts.length) {
-			setSelectedConflictId(null);
-			return;
-		}
-
-		if (
-			!selectedConflictId ||
-			!conflicts.some((item) => item._id === selectedConflictId)
-		) {
-			setSelectedConflictId(conflicts[0]._id);
-		}
-	}, [conflicts, selectedConflictId]);
-
-	const selectedConflict = useMemo(
-		() => conflicts.find((item) => item._id === selectedConflictId) ?? null,
-		[conflicts, selectedConflictId]
-	);
-
-	const updateConflictMutation = useMutation({
-		mutationFn: ({
-			actionTaken,
-			conflictId,
-			status,
-		}: {
-			actionTaken?: string;
-			conflictId: string;
-			status: Extract<ConflictStatus, 'DISMISSED' | 'RESOLVED'>;
-		}) => updateConflict({ actionTaken, conflictId, status, therapistId }),
-		onSuccess: () => {
-			queryClient.invalidateQueries({ queryKey: ['conflicts', therapistId] });
-			queryClient.invalidateQueries({
-				queryKey: ['conflictCount', therapistId],
-			});
-		},
-		onError: () => {
-			showAlert({
-				message: t('conflicts.actions.error'),
-				severity: 'error',
-			});
-		},
-	});
-
-	const handleUpdateConflict = useCallback(
-		(input: {
-			actionTaken?: string;
-			conflictId: string;
-			status: Extract<ConflictStatus, 'DISMISSED' | 'RESOLVED'>;
-		}) => {
-			if (useMockConflicts) {
-				return;
-			}
-
-			updateConflictMutation.mutate(input);
-		},
-		[updateConflictMutation, useMockConflicts]
-	);
+	const {
+		conflicts,
+		handleUpdateConflict,
+		isLoading,
+		isUpdating,
+		selectedConflict,
+		selectedConflictId,
+		setSelectedConflictId,
+		setStatusFilter,
+		setTypeFilter,
+		statusFilter,
+		typeFilter,
+	} = useConflictsPageState({ t });
 
 	return (
 		<PageLayout
 			title={t('conflicts.title')}
-			subTitle={
-				useMockConflicts
-					? `${t('conflicts.subtitle')} ${t('conflicts.mock-mode')}`
-					: t('conflicts.subtitle')
-			}
-			isLoading={!useMockConflicts && isLoading}
+			subTitle={t('conflicts.subtitle')}
+			isLoading={isLoading}
 		>
 			<ConflictsLayout>
 				<ConflictsSidebar>
@@ -164,12 +63,6 @@ export const ConflictsPage = () => {
 						</SidebarTitleRow>
 						<SidebarSubtitle>{t('conflicts.queue.subtitle')}</SidebarSubtitle>
 					</SidebarHeader>
-
-					{useMockConflicts ? (
-						<MockModeNotice>
-							<MockModeText>{t('conflicts.mock-notice')}</MockModeText>
-						</MockModeNotice>
-					) : null}
 
 					<FiltersSection>
 						<FiltersLabel>{t('conflicts.filters.status')}</FiltersLabel>
@@ -248,7 +141,7 @@ export const ConflictsPage = () => {
 
 				<ConflictDetail
 					conflict={selectedConflict}
-					isUpdating={updateConflictMutation.isPending}
+					isUpdating={isUpdating}
 					onUpdateConflict={handleUpdateConflict}
 				/>
 			</ConflictsLayout>
