+++ src/pages/conflicts/components/conflict-detail/ConflictDetail.types.ts
@@ -1,14 +1,9 @@
-import type {
-	ConflictStatus,
-	IConflict,
-} from '@psycron/api/user/conflicts/index.types';
+import type { IConflict } from '@psycron/api/user/conflicts/index.types';
+
+import type { UpdateConflictInput } from '../../hooks/useConflictsPageState';
 
 export interface ConflictDetailProps {
 	conflict: IConflict | null;
 	isUpdating: boolean;
-	onUpdateConflict: (input: {
-		actionTaken?: string;
-		conflictId: string;
-		status: Extract<ConflictStatus, 'DISMISSED' | 'RESOLVED'>;
-	}) => void;
+	onUpdateConflict: (input: UpdateConflictInput) => void;
 }
